+++ src/test/java/io/github/astrapi69/export/excel/poi/ExportExcelExtensionsTest.java
@@ -62,8 +62,8 @@ import io.github.astrapi69.lang.ClassExtensions;
 public class ExportExcelExtensionsTest
 {
 	final String twoDimArray[][] = { { "1", "a", "!" }, { "2", "b", "?" }, { "3", "c", "%" } };
-	final String twoDimArrayDouble[][] = { { "1.0", "a", "!" }, { "2.0", "b", "?" },
-			{ "3.0", "c", "%" } };
+	final String[][] twoDimArrayDouble = { { "1", "a", "!" }, { "2", "b", "?" },
+			{ "3", "c", "%" } };
 	File emptyWorkbook;
 	Workbook workbook;